+++ packages/core/CHANGELOG.md
@@ -121,9 +121,11 @@ All notable changes to this project will be documented in this file. See
 
 ### Bug Fixes
 
--   fix pie chart alignment
-    ([#716](https://github.com/carbon-design-system/carbon-charts/issues/716))
-    ([638c331](https://github.com/carbon-design-system/carbon-charts/commit/638c331b7fa8fbc73cf557cf99a6b3af2435b492))
+* fix pie chart alignment ([#716](https://github.com/carbon-design-system/carbon-charts/issues/716)) ([638c331](https://github.com/carbon-design-system/carbon-charts/commit/638c331b7fa8fbc73cf557cf99a6b3af2435b492))
+
+
+
+
 
 # Change Log